requirements.txt
scikit-image==0.18.1
numpy==1.20.3
scipy==1.6.3
Django==2.2.19
django-environ==0.4.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
environ==1.0
pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
celery==5.0.5
django-celery-results==2.0.1
pyyaml==5.4.1
uritemplate==3.0.1
djangorestframework-simplejwt==4.6.0